Refactor rate limiting implementation: move to pkg/bot/ratelimit and update imports
